tests/wpt/tests/xhr/open-url-about-blank-window.htm

<!doctype html>
<html>
  <head>
    <title>XMLHttpRequest: open() resolving URLs (about:blank iframe)</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <link rel="help" href="https://xhr.spec.whatwg.org/#the-open()-method" data-tested-assertations="following::ol/li[2]/ol/li[2] following::ol/li[7] following::ol/li[14]/ul/li[2]" />
    <link rel="help" href="https://xhr.spec.whatwg.org/#the-responsetext-attribute" data-tested-assertations="following::ol/li[4]" />
    <link rel="help" href="https://xhr.spec.whatwg.org/#concept-xmlhttprequest-document" data-tested-assertations=".." />
  </head>
  <body>
    <div id="log"></div>
    <iframe src="about:blank"></iframe>
    <script>
      test(function() {
        var client = new self[0].XMLHttpRequest()
        client.open("GET", "folder.txt", false)
        client.send("")
        assert_equals(client.responseText, "top\n")
      })
    </script>
  </body>
</html>
